VolumePanel: Fix Theme Engine support with On-The-Spot dialog
Our on the spot feature needs to write to system. Since VolumePanel when is
recreated uses theme engine context (and in this case is fked), pass to on the spot dialog always the
first system initialized one. Thanks to @David96 helping find this out :)

This way we lose theming for the safe headset on the spot dialog,
but we're already finding a really good fix, we have something already :)

Fix AOSPA-482

Signed-off-by: Carlo Savignano <stevewatersy@gmail.com>
Change-Id: I492cd8e24b09d856d23697b84113664379703cf1
Conflicts:
	core/java/android/view/VolumePanel.java

Conflicts:
	media/java/android/media/AudioService.java